refactor: centralize prompt manager module
