+++ tests/behavior/steps/wsde_edrr_integration_steps.py
@@ -22,7 +22,7 @@ from devsynth.domain.models.wsde import WSDETeam
 from devsynth.application.memory.memory_manager import MemoryManager
 from devsynth.application.code_analysis.analyzer import CodeAnalyzer
 from devsynth.application.code_analysis.ast_transformer import AstTransformer
-from devsynth.application.prompts.prompt_manager import PromptManager
+from devsynth.application.requirements.prompt_manager import PromptManager
 from devsynth.application.documentation.documentation_manager import DocumentationManager
 from devsynth.methodology.base import Phase
 from devsynth.domain.models.agent import AgentConfig, AgentType
